% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcpr-package.R
\docType{package}
\name{MCPR-package}
\alias{MCPR}
\alias{MCPR-package}
\title{MCPR Package}
\description{
Configures platform-specific socket URLs when package loads into R session
}
\details{
Model Context Protocol for R
}
\seealso{
Useful links:
\itemize{
  \item \url{https://phisanti.github.io/MCPR/}
  \item \url{https://github.com/phisanti/MCPR}
  \item Report bugs at \url{https://github.com/phisanti/MCPR/issues}
}

}
\author{
\strong{Maintainer}: Santiago Cano-Muniz \email{tisalon@outlook.com}

}
\keyword{internal}
